fix: Atualizando e Adicionando as classes novas

+++ src/main/java/br/com/ifba/conectaedu/repository/ExameNacionalRepository.java
@@ -2,9 +2,8 @@ package br.com.ifba.conectaedu.repository;
 
 import br.com.ifba.conectaedu.entity.ExameNacional;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
 
-public interface ExameNacionalRepository extends JpaRepository<ExameNacional, Long> {
+public interface ExameNacionalRepository extends JpaRepository<ExameNacional, Integer> {
 
 
 }
